add new API for specifying default animation
